Add HTTP user agent.
